update to mantine ui
